+++ HISTORY.md
@@ -114,3 +114,8 @@
 ## 0.4.7
 - export the functions to be used in other apps
 - add `antd` theme 
+
+## 0.4.8
+- fix a bug `window: not defined` error in `getCurrentChainId`
+- add `getIpfsDetails`
+- add `<Loader />` component